[WIP] adds Guidelines
Closes #42

This PR handles adding guideline resources are described in #42.

To follow the requested URL structure (comment:
https://github.com/panvol/pandemic-volunteers/issues/42#issuecomment-615248137),
an default locale of `en` is introduced and used as a scope in routes,
to ensure that `/en/guidelines` and `/guidelines` route default to the
same resource

Because `guidelines` are ephemeral, a presenter is used to display
initial guidelines that have guideline categories that, in turn, have
many articles. Presenters also take care of fetching and rendering
metadata stored in YML files

TODO:
- add markdown parsing / rendering based on https://github.com/panvol/pandemic-volunteers/pull/44/files#diff-e55a6c615d18aeb75e0610b835b16658
- make methods such as `GuidelineCategory.extract_slug` etc., metadata
  file fetching and loading, and loading dirs via `Pathname` reusable.
  This is because GuidelineCategory and GuidelineArticle will be sharing
  90% of their methods. GuidelineArticle will also need to be able to
  load its own metadata
- move `GuidelinesPresenter` to `app/models` since `GuidelineCategory`
  is very similar and acts as a stand-in for a model
- add designs